Tests: Move PowerShell get_credentials() tests
Move the tests for the PowerShell exploiter's get_credentials() function
to test_credential_generators.py, since get_credentials() is now
contained in credential_generators.py
